fixed warning about different signess
git-svn-id: https://btanks.svn.sourceforge.net/svnroot/btanks/trunk@4756 5c0bf991-d514-0410-a65f-ce4832e571f3

+++ net/monitor.cpp
@@ -267,7 +267,7 @@ TRY {
 					if (t->size_task) {
 						unsigned long len = ntohl(*((uint32_t *)(t->data->getPtr())));
 						GET_CONFIG_VALUE("multiplayer.maximum-packet-length", int, max_len, 1024 * 1024);
-						if (len > max_len)
+						if (len > (unsigned long)max_len)
 							throw_ex(("recv'ed packet length of %u. it seems to be far too long for regular packet (probably broken/obsoleted client)", (unsigned int)len));
 						unsigned char flags = *((unsigned char *)(t->data->getPtr()) + 4);
 						//LOG_DEBUG(("added task for %u bytes. flags = %02x", len, flags));
